develop/requirements.txt
acicobra===4.2-3h
acimodel===4.2-3h
acitoolkit==0.4
aiohttp==3.6.2
alembic==1.4.0
ansible==2.9.4
appdirs==1.4.3
arrow==0.15.5
arya==1.1.5
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
bcrypt==3.1.7
binaryornot==0.4.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
ciscoconfparse==1.4.11
ciscosparkapi==0.9.2
Click==7.0
cliff==2.18.0
cmake==3.16.3
cmd2==0.8.9
colorama==0.4.3
cookiecutter==1.7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
debtcollector==1.22.0
decorator==4.4.1
deepdiff==4.2.0
dill==0.3.1.1
distro==1.4.0
Django==2.2.10
django-cacheops==4.2
django-cors-headers==3.2.1
django-debug-toolbar==2.1
django-filter==2.2.0
django-js-asset==1.2.2
django-mptt==0.9.1
django-prometheus==1.1.0
django-rq==2.2.0
django-tables2==2.2.1
django-taggit==1.2.0
django-taggit-serializer==0.1.7
django-timezone-field==4.0
djangorestframework==3.10.3
dnspython==1.16.0
dogpile.cache==0.9.0
dominate==2.4.0
drf-yasg==1.17.0
extras==1.0.0
fixtures==3.0.0
Flask==1.1.1
Flask-Admin==1.5.5
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
funcy==1.14
future==0.18.2
genie==20.1
genie.libs.conf==20.1
genie.libs.filetransferutils==20.1
genie.libs.ops==20.1
genie.libs.parser==20.1
genie.libs.robot==20.1
genie.libs.sdk==20.1
gitdb2==2.0.6
GitPython==3.0.7
graphviz==0.13.2
idna==2.6
importlib-metadata==1.5.0
inflection==0.3.1
ipaddress==1.0.23
iso8601==0.1.12
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.3
jinja2-time==0.2.0
jmespath==0.9.4
jsonpatch==1.25
jsonpickle==1.3
jsonpointer==2.0
jsonschema==3.2.0
junit-xml==1.9
junos-eznc==2.2.1
keystoneauth1==3.18.0
lazy-object-proxy==1.4.3
linecache2==1.0.0
lxml==4.5.0
Mako==1.1.1
Markdown==2.6.11
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==8.2.0
msgpack==0.6.2
multidict==4.7.5
munch==2.5.0
mypy-extensions==0.4.3
napalm==2.5.0
ncclient==0.6.7
netaddr==0.7.19
netifaces==0.10.9
netmiko==2.4.2
nornir==2.4.0
nxapi-plumbing==0.5.2
openstacksdk==0.40.0
ordered-set==3.1.1
os-client-config==2.0.0
os-service-types==1.7.0
osc-lib==2.0.0
oslo.config==7.0.0
oslo.context==3.0.0
oslo.db==6.0.0
oslo.i18n==3.25.1
oslo.log==4.0.0
oslo.serialization==2.29.2
oslo.utils==3.42.1
packaging==20.1
paramiko==2.7.1
passlib==1.7.2
pathspec==0.7.0
pbr==5.4.4
pika==1.1.0
Pillow==7.0.0
pluggy==0.13.1
ply==3.11
poyo==0.5.0
prettytable==0.7.2
prometheus-client==0.7.1
psutil==5.7.0
psycopg2-binary==2.8.4
py==1.8.1
py-gfm==0.1.4
py-radix==0.10.0
pyang==1.7.4
pyats==20.1
pyats.aereport==20.1
pyats.aetest==20.1
pyats.async==20.1
pyats.connections==20.1
pyats.datastructures==20.1
pyats.easypy==20.1
pyats.kleenex==20.1
pyats.log==20.1
pyats.reporter==20.1
pyats.results==20.1
pyats.robot==20.1
pyats.tcl==20.1
pyats.topology==20.1.1
pyats.utils==20.1
pybind11==2.4.3
pycparser==2.19
pycryptodome==3.9.4
pyeapi==0.8.3
pygraphviz==1.5
pyIOSXR==0.53
pylint==2.4.4
PyMySQL==0.9.3
PyNaCl==1.3.0
pynetbox==4.2.5
pyparsing==2.4.6
pyperclip==1.7.0
pyrsistent==0.15.7
pyserial==3.4
pytest==5.3.5
python-dateutil==2.8.1
python-editor==1.0.4
python-keystoneclient==3.22.0
python-mimeparse==1.6.0
python-neutronclient==6.14.0
pytz==2019.3
PyYAML==5.3
redis==3.3.11
requests==2.22.0
requests-toolbelt==0.9.1
requestsexceptions==1.4.0
rfc3986==1.3.2
robotframework==3.1.2
rq==1.2.2
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scp==0.13.2
simplejson==3.17.0
six==1.14.0
smmap2==2.0.5
SQLAlchemy==1.3.13
sqlalchemy-migrate==0.13.0
sqlparse==0.3.0
stevedore==1.31.0
svgwrite==1.3.1
swagger-spec-validator==2.4.3
tabulate==0.8.6
Tempita==0.5.2
testresources==2.0.1
testscenarios==0.5.0
testtools==2.3.0
textfsm==1.1.0
tqdm==4.43.0
traceback2==1.4.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
ucsmsdk==0.9.9
unicon==20.1
unicon.plugins==20.1.1
unittest2==1.1.0
uritemplate==3.0.1
urllib3==1.22
visitor==0.1.3
wcwidth==0.1.8
WebOb==1.8.6
websocket-client==0.57.0
Werkzeug==1.0.0
whichcraft==0.6.1
wrapt==1.11.2
WTForms==2.2.1
xmltodict==0.11.0
yamllint==1.20.0
yarl==1.4.2
zipp==2.2.0
